+++ sqlparser/test/analysis/errors/syntax_error_test.dart
@@ -1,5 +1,6 @@
 // tests for syntax errors revealed during static analysis.
 
+import 'package:sqlparser/sqlparser.dart';
 import 'package:sqlparser/src/analysis/analysis.dart';
 import 'package:sqlparser/src/engine/sql_engine.dart';
 import 'package:test/test.dart';
@@ -17,6 +18,27 @@ void main() {
           .having((e) => e.type, 'type', AnalysisErrorType.synctactic)
           .having((e) => e.message, 'message',
               contains('Expected a conflict clause'))
+    ]);
+  });
+
+  test('complex expressions in PRIMARY KEY clause', () {
+    final engine = SqlEngine();
+    final parseResult = engine.parse('''
+      CREATE TABLE tbl (
+        foo INTEGER NOT NULL,
+        bar TEXT NOT NULL,
+        PRIMARY KEY (foo DESC, bar || 'test')
+      );
+    ''');
+    engine.registerTable(const SchemaFromCreateTable()
+        .read(parseResult.rootNode as CreateTableStatement));
+    final analyzeResult = engine.analyzeParsed(parseResult);
+
+    expect(analyzeResult.errors, [
+      const TypeMatcher<AnalysisError>()
+          .having((e) => e.type, 'type', AnalysisErrorType.synctactic)
+          .having((e) => e.message, 'message',
+              contains('Only column names can be used in a PRIMARY KEY clause'))
     ]);
   });
 }